requirements.txt
Django==1.10
django-markdown-deux==1.0.5

